moved the owner commands into a separate cog
